Virtual methods: Fixed error & clarified
The original text stated that "When a virtual method is invoked, the *run-time type* of the instance ... determines the actual method implementation ...", which is not entirely correct, because it is only possible but not ensured. The derived class needs to override that virtual method explicitly in order to ensure that the *run-time type* will determine the actually invoked implementation. Without the override keyword, a virtual method will behave similarly to a non-virtual method.